feat: support expression statements and debugging abilities for ast
In this language, we want to be able to treat standalone Expressions as
Statements (e.g.: "x + 10;"). Therefore we created a type that
implements the Statement interface called ExpressionStatement.

In addition to this, we fledged out the "ast" package in order to have
the ability to debug and test any of our types that implement the
Statement interface (ReturnStatement, LetStatement, ExpressionStatement,
etc). This means that we can initialize and populate our
Program.Statements and then get our whole program back as a string with
the String() method. Also created tests to check if this functionality
works properly.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>