feat: implement Pratt Parser (expression parser)
In a Pratt Parser, "semantic code" refers to the association of parsing
functions with token types. This means that whenever a token type is
encountered, the parsing functions are called to parse the appropriate
expression and return an ast.Expression. We can have up to two parsing
functions that parse the Expression (prefix or infix positional parsing
function).

Signed-off-by: Sam Gabel <sam@gabelcloud.com>