feat: parse standalone identifiers as expression statements
Identifiers are possibly the simplest expression type. Identifiers can
be used in other contexts as well, like arguments in a function call, as
operands in an infix expression, etc. We want to start small and first
handle standalone identifiers as expression statements (e.g.:
"foobar;").

We also need to lay the groundwork for operator precedence, this can be
seen as order of operations, or which operation do we want to evaluate
first.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>