feat: parse prefix operators in prefix expressions
In this language we will only have 2 prefix operators which are "!" and
"-". We have to make sure that we evaluate the expression that comes
after the prefix operator as well. In this case we will have a prefix
operator and the left expression. It should also be noted that anything
that is considered to be an expression can follow a prefix operator.

Signed-off-by: Sam Gabel <sam@gabelcloud.com>